backend/requirements.txt
django>=4.2
djangorestframework>=3.14
mysqlclient>=2.2
django-cors-headers>=4.9
